LayoutTests/fast/events/touch/ios/slider-touch-drag-small-thumb.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
    <meta name="viewport" content="width=device-width">
    <script src="../../../resources/js-test.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <style>
        input[type="range"] {
            width: 300px;
            height: 6px;
            background: black;
        }

        input[type="range"]::-webkit-slider-thumb {
            appearance: none;
            background: blue;
            width: 6px;
            height: 6px;
        }
    </style>
    <script>
        jsTestIsAsync = true;
        var inputFired = false;

        addEventListener("load", async () => {
            description("Tests that dragging a range input with a small custom thumb fires input events");

            if (!window.testRunner || !testRunner.runUIScript) {
                debug("This test requires WebKitTestRunner.");
                finishJSTest();
                return;
            }

            const slider = document.getElementById("slider");
            const rect = slider.getBoundingClientRect();

            const startX = rect.left + rect.width / 2;
            const startY = rect.top + rect.height / 2;
            const endX = startX + 100;

            slider.addEventListener("input", () => { inputFired = true; }, { once: true });

            await UIHelper.dragFromPointToPoint(startX, startY, endX, startY, 0.5);

            shouldBeTrue("inputFired");
            finishJSTest();
        });
    </script>
</head>
<body>
<input id="slider" type="range" min="0" max="100" value="50">
<div id="description"></div>
<div id="console"></div>
</body>
</html>
