LayoutTests/fast/events/touch/ios/slider-touch-drag-small-thumb-expected.txt


Tests that dragging a range input with a small custom thumb fires input events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS inputFired is true
PASS successfullyParsed is true

TEST COMPLETE
